[team01] 너무 많이 바꿨는데 author: onelight detail: 내가 맡은 부분 거의 다 수정 + 시스템/커뮤니티 구분
