+++ teamBootstrap/src/main/resources/templates/contract/addBuyer.html
@@ -19,42 +19,42 @@ layout:decorate="@{layout/default_contract}">
                   
                   <div class="form-group ">
                     <label for="firstname" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">회사명</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class=" form-control" id="clientAccountName" name="clientAccountName" type="text">
                     </div>
                   </div>
                   
                   <div class="form-group ">
                     <label for="lastname" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">전화번호</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class=" form-control" id="clientAccountPhone" name="clientAccountPhone" type="text">
                     </div>
                   </div>
                   
                   <div class="form-group ">
                     <label for="email" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">대표자</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class="form-control " id="clientAccountCeo" name="clientAccountCeo" type="text">
                     </div>
                   </div>
                   
                   <div class="form-group ">
                     <label for="email" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">담당자</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class="form-control " id="clientAccountManager" name="clientAccountManager" type="text">
                     </div>
                   </div>
                   
                   <div class="form-group ">
                     <label for="email" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">사업자번호</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class="form-control " id="clientAccountPhone" name="clientAccountPhone" type="text">
                     </div>
                   </div>
                   
                   <div class="form-group ">
                     <label for="email" class="control-label col-lg-2"><font style="vertical-align: inherit;"><font style="vertical-align: inherit;">주소</font></font></label>
-                    <div class="col-lg-10">
+                    <div class="col-lg-6">
                       <input class="form-control " id="clientAccountAddr" name="clientAccountAddr" type="text">
                     </div>
                   </div>
