+++ teamBootstrap/src/main/resources/templates/fragments/header.html
@@ -32,6 +32,9 @@
         <a th:href="@{/humanResources}">
         	<span class="menu">인사관리</span>
         </a>
+        <a th:href="@{/community}">
+        	<span class="menu">커뮤니티</span>
+        </a>
         <a th:href="@{/system}">
         	<span class="menu">시스템관리</span>
         </a>
