+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_rawMaterials.html
@@ -32,9 +32,10 @@
               <span>자재출고  #[한빛]</span>
               </a>
             <ul class="sub">
-              <li><a th:href="@{/addExWarehousing}">의뢰별 소요자재 등록  #[한빛]</a></li>
-              <li><a th:href="@{/exWarehousingList}">의뢰별 소요자재 조회  #[한빛]</a></li>
-              <li><a th:href="@{/exWarehousingStatus}">출고현황  #[한빛]</a></li>
+              <li><a th:href="@{/addMaterialsUse}">의뢰별 소요자재 등록  #[한빛]</a></li>
+              <li><a th:href="@{/materialsUseList}">의뢰별 소요자재 조회  #[한빛]</a></li>
+              <li><a th:href="@{/addExWarehousing}">출고등록  #[한빛]</a></li>
+              <li><a th:href="@{/exWarehousingList}">출고현황  #[한빛]</a></li>
             </ul>
 
         </ul>
