+++ teamBootstrap/src/main/resources/templates/fragments/sidebar_system.html
@@ -16,28 +16,7 @@
             </ul>
           </li>
           
-          <li class="sub-menu">
-            <a href="javascript:;">
-              <i class="fa fa-book"></i>
-              <span>커뮤니티</span>
-              </a>
-            <ul class="sub">
-              <li><a th:href="@{/lobby}">로비</a></li>
-              <li><a th:href="@{/messenger}">메신저</a></li>
-            </ul>
-          </li>
-          
-          <li class="sub-menu">
-            <a href="javascript:;">
-              <i class="fa fa-book"></i>
-              <span>Q&A</span>
-              </a>
-            <ul class="sub">
-              <li><a th:href="@{/qAndA}">자주하는 질문</a></li>
-            </ul>
-          </li>
  
-    
         </ul>
         <!-- sidebar menu end-->
       </div>
