+++ teamBootstrap/src/main/resources/templates/humanResources/humanResources.html
@@ -68,6 +68,7 @@ layout:decorate="@{layout/default_humanResources}">
                     <th><i class=" fa fa-bookmark"></i>담당직책</th>
                     <th><i class=" fa fa-bookmark"></i>이메일</th>
                     <th><i class=" fa fa-bookmark"></i>전화번호</th>
+                    <th><i class=" fa fa-bookmark"></i>등록일자</th>
                     <th><i class=" fa fa-pencil"></i>수정</th>
                   </tr>
                 </thead>
@@ -83,11 +84,13 @@ layout:decorate="@{layout/default_humanResources}">
                     <td>onelight@naver.com</td>
                     <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
                     <td>010-7170-7123</td>
-                    <td> 
-                      
-                      <button class="btn btn-primary btn-xs"><i class="fa fa-pencil"></i></button>
-                      
-                    </td>
+                    <td>2021.06.22 14:20</td>
+                  	<td>  
+                  		<button type="submit" th:href="@{/modifyHumanResources}" class="btn btn-primary btn-xs">
+                  			<i class="fa fa-pencil"></i>
+                  				수정
+                  		</button>
+                  	</td>
                   </tr>
                   
                
@@ -99,7 +102,7 @@ layout:decorate="@{layout/default_humanResources}">
             
 		        <div class="form-group text-right">
 		          <div class="col-lg-offset-2 col-lg-10">
-		           <a th:href="@{/addHumanResources}" class="btn btn-success btn-xs"><i class="fa fa-check"></i>등록</a>
+		           <a th:href="@{/addHumanResources}" class="btn btn-success btn-xs"><i class="fa fa-check"></i>사원등록</a>
 		           <a id="deleteBtn" class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i>삭제</a>
 		          </div>
 		        </div>
