+++ teamBootstrap/src/main/resources/templates/humanResources/humanResourcesList.html
@@ -85,11 +85,12 @@ layout:decorate="@{layout/default_humanResources}">
                     <!-- 날짜형태로 받을 수 있게 설정해야함 th:datetime -->
                     <td>010-7170-7123</td>
                     <td>2021.06.22 14:20</td>
-                    <td> 
-                      
-                      <button class="btn btn-primary btn-xs"><i class="fa fa-pencil"></i></button>
-                      
-                    </td>
+                  	<td>  
+                  		<button type="submit" th:href="@{/modifyHumanResources}" class="btn btn-primary btn-xs">
+                  			<i class="fa fa-pencil"></i>
+                  				수정
+                  		</button>
+                  	</td>
                   </tr>
