+++ teamBootstrap/src/main/resources/templates/humanResources/modifyHumanResources.html
@@ -15,7 +15,7 @@ layout:decorate="@{layout/default_humanResources}">
 	<div class="form-panel">
 	 <h4 class="mb"><i class="fa fa-angle-right"></i><font style="vertical-align: inherit;"><font style="vertical-align: inherit;"> 사원등록</font></font></h4>
               <div class="form">
-                <form class="cmxform form-horizontal style-form" id="signupForm" method="post"  th:action="@{/addHumanResources}">
+                <form class="cmxform form-horizontal style-form" id="signupForm" method="post"  th:action="@{/modifyHumanResources}">
                   <div class="form-group">
                   <label class="col-sm-2 col-sm-2 control-label" style="font-weight: 600;">권한</label>
                   <div class="col-sm-6">
