+++ teamBootstrap/src/main/resources/templates/layout/default_community.html
@@ -1,0 +1,134 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta name="description" content="">
+  <meta name="author" content="Dashboard">
+  <meta name="keyword" content="Dashboard, Bootstrap, Admin, Template, Theme, Responsive, Fluid, Retina">
+  <title>Dashio - Bootstrap Admin Template</title>
+
+  <!-- Favicons -->
+  <link th:href="@{Dashio/img/favicon.png}" rel="icon">
+  <link th:href="@{Dashio/img/apple-touch-icon.png}" rel="apple-touch-icon">
+
+  <!-- Bootstrap core CSS -->
+  <link th:href="@{Dashio/lib/bootstrap/css/bootstrap.min.css}" rel="stylesheet">
+  <!--external css-->
+  <link th:href="@{Dashio/lib/font-awesome/css/font-awesome.css}" rel="stylesheet" />
+  <link rel="stylesheet" type="text/css" th:href="@{Dashio/css/zabuto_calendar.css}">
+  <link rel="stylesheet" type="text/css" th:href="@{Dashio/lib/gritter/css/jquery.gritter.css}" />
+  <!-- Custom styles for this template -->
+  <link th:href="@{Dashio/css/style.css}" rel="stylesheet">
+  <link th:href="@{Dashio/css/style-responsive.css}" rel="stylesheet">
+  <script th:src="@{Dashio/lib/chart-master/Chart.js}"></script>
+
+</head>
+
+<body>
+  <section id="container">
+		<header th:replace="/fragments/header::header"></header>
+
+    <aside>
+		<div th:replace="/fragments/sidebar_community::sidebar_community"></div>
+    </aside>
+    <!--main content start-->
+    <section id="main-content">
+      <section class="wrapper">
+       
+        <div class="row">
+          <div class="col-lg-9 main-chart">
+            <!--CUSTOM CHART START -->
+          <th:block layout:fragment="pageTitle"></th:block>
+				
+          <th:block layout:fragment="container"></th:block>
+          </div>
+          <!-- !!!!!!본문!!!!!!! -->
+          		
+        </div>
+        <!-- /row -->
+      </section>
+    </section>
+	   	<footer th:replace="/fragments/footer"></footer>
+
+  </section>
+ <script th:src="@{Dashio/lib/jquery/jquery.min.js}"></script>
+  <script th:src="@{Dashio/lib/bootstrap/js/bootstrap.min.js}"></script>
+  
+  <script class="include" type="text/javascript" th:src="@{Dashio/lib/jquery.dcjqaccordion.2.7.js}"></script>
+  
+  <script th:src="@{Dashio/lib/jquery.scrollTo.min.js}"></script>
+  <script th:src="@{Dashio/lib/jquery.nicescroll.js}" type="text/javascript"></script>
+  <script th:src="@{Dashio/lib/jquery.sparkline.js}"></script>
+  <!--common script for all pages-->
+  <script th:src="@{Dashio/lib/common-scripts.js}"></script>
+  <script type="text/javascript" th:src="@{Dashio/lib/gritter/js/jquery.gritter.js}"></script>
+  <script type="text/javascript" th:src="@{Dashio/lib/gritter-conf.js}"></script>
+  <!--script for this page-->
+  <script th:src="@{Dashio/lib/sparkline-chart.js}"></script>
+  <script th:src="@{Dashio/lib/zabuto_calendar.js}"></script>
+  <script type="text/javascript">
+    $(document).ready(function() {
+      var unique_id = $.gritter.add({
+        // (string | mandatory) the heading of the notification
+        title: 'Welcome to Dashio!',
+        // (string | mandatory) the text inside the notification
+        text: 'Hover me to enable the Close Button. You can hide the left sidebar clicking on the button next to the logo.',
+        // (string | optional) the image to display on the left
+        image: 'img/ui-sam.jpg',
+        // (bool | optional) if you want it to fade out on its own or just sit there
+        sticky: false,
+        // (int | optional) the time you want it to be alive for before fading out
+        time: 8000,
+        // (string | optional) the class name you want to apply to that specific message
+        class_name: 'my-sticky-class'
+      });
+
+      return false;
+    });
+  </script>
+  <script type="application/javascript">
+    $(document).ready(function() {
+      $("#date-popover").popover({
+        html: true,
+        trigger: "manual"
+      });
+      $("#date-popover").hide();
+      $("#date-popover").click(function(e) {
+        $(this).hide();
+      });
+
+      $("#my-calendar").zabuto_calendar({
+        action: function() {
+          return myDateFunction(this.id, false);
+        },
+        action_nav: function() {
+          return myNavFunction(this.id);
+        },
+        ajax: {
+          url: "show_data.php?action=1",
+          modal: true
+        },
+        legend: [{
+            type: "text",
+            label: "Special event",
+            badge: "00"
+          },
+          {
+            type: "block",
+            label: "Regular event",
+          }
+        ]
+      });
+    });
+
+    function myNavFunction(id) {
+      $("#date-popover").hide();
+      var nav = $("#" + id).data("navigation");
+      var to = $("#" + id).data("to");
+      console.log('nav ' + nav + ' to: ' + to.month + '/' + to.year);
+    }
+  </script>
+</body>
+
+</html>
