+++ teamBootstrap/src/main/resources/templates/rawMaterials/exWarehousingList.html
@@ -5,7 +5,7 @@ layout:decorate="@{layout/default_rawMaterials}">
 
 <th:block layout:fragment="pageTitle">
 	   <div class="border-head"> 
-	    <h3 th:text="${title}"></h3>
+	    <h3>자재관리</h3>
 	    </div>
 </th:block>
 
@@ -22,19 +22,21 @@ layout:decorate="@{layout/default_rawMaterials}">
 </th:block>
 
 <th:block layout:fragment="container">
-자재관리
-	
 	
 	<div class="row mt">
      <div class="col-md-12">
-          <h4><i class="fa fa-truck"></i>의뢰별 소요자재 목록/조회</h4>
+          <h4><i class="fa fa-truck"></i>자재 출고 현황 목록/조회</h4>
           <hr>  
 			<div class="form-group" style="padding-left: 50px">
 				<div class="col-md-3">
 				 <select class="form-control">
+				   <option>원부자재 코드</option>
 				   <option>재료명</option>
 				   <option>색상명</option>
-
+				   <option>수량</option>
+				   <option>단위</option>
+				   <option>출고날짜</option>
+				   <option>등록일시</option>
 				  </select>
 				</div>
 				<div class="col-md-6">
@@ -42,7 +44,7 @@ layout:decorate="@{layout/default_rawMaterials}">
 				</div>
 				<div class="col-md-2">
 				  <!-- 버튼 누르면 검색할 수 있는 창으로 연결 -->            
-				  <button type="button" class="btn btn-primary">검색</button>
+				  <button type="button" class="btn btn-primary">검색하여 가져오기</button>
 				</div>
 	       </div>
 	       
@@ -59,58 +61,66 @@ layout:decorate="@{layout/default_rawMaterials}">
                   <tr>
                     <th><input type="checkbox" class="checked"></th>
                     <th><i class="fa fa-bullhorn"></i>번호</th>
-                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>원부자재코드</th>
-                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>원부자재 발주코드</th>
-                    <th><i class="fa fa-bookmark"></i>재료명</th>
+                    <th class="hidden-phone"><i class="fa fa-bookmark"></i>원부자재 코드</th>
+                    <th><i class="fa fa-bookmark"></i>원부자재 발주코드</th>
+                    <th><i class=" fa fa-bookmark"></i>재료명</th>
                     <th><i class=" fa fa-bookmark"></i>색상명</th>
                     <th><i class=" fa fa-bookmark"></i>특징</th>
                     <th><i class=" fa fa-bookmark"></i>수량</th>
                     <th><i class=" fa fa-bookmark"></i>단위</th>
                     <th><i class=" fa fa-bookmark"></i>출고일자</th>
+                    <th><i class=" fa fa-bookmark"></i>출고현황</th>
                     <th><i class=" fa fa-bookmark"></i>등록일자</th>
                     <th><i class=" fa fa-pencil"></i>수정</th>
                   </tr>
                 </thead>
                 <tbody>
                   <tr>
-                  	<td><input type="checkbox" class="checked"></td>
+                    <th><input type="checkbox" class="checked"></th>
                     <td class="hidden-phone">1</td>
-                    <td class="hidden-phone">rw001</td>      
-                    <td class="hidden-phone">rwpw001</td>
-                    <td>원색실</td>
+                    <td class="hidden-phone">rw001</td>
+                    <td class="hidden-phone">rwex001</td>
+                    <td class="hidden-phone">원단실</td>
                     <td>개나리색</td>
                     <td>해당없음</td>
-                    <td>100</td>
+                    <td>200</td>
                     <td>개</td>
-                    <td>2021.06.22 11:53</td>
-                    <td>2021.06.22 11:53</td>
-                    <td> 
-                      
-                      <button class="btn btn-primary btn-xs"><i class="fa fa-pencil"></i></button>
-                      
-                    </td>
-                  </tr>
+                    <td>2021-01-18 16:30</td>             
+                    <td><span class="label label-info label-mini">출고완료</span></td>
+                    <td>2021-01-18 16:30</td>                
+                  	<td>  
+                  		<button type="submit" th:href="@{/modifyHumanResources}" class="btn btn-primary btn-xs">
+                  			<i class="fa fa-pencil"></i>
+                  				수정
+                  		</button>
+                  	</td>
+                   </tr>
                   <tr>
-                  	<td><input type="checkbox" class="checked"></td>
+                    <th><input type="checkbox" class="checked"></th>
+                    
                     <td class="hidden-phone">1</td>
-                    <td class="hidden-phone">rw001</td>      
-                    <td class="hidden-phone">rwpw001</td>
-                    <td>원색실</td>
+                    <td class="hidden-phone">rw001</td>
+                    <td class="hidden-phone">rwex001</td>
+                    <td class="hidden-phone">원단실</td>
                     <td>개나리색</td>
                     <td>해당없음</td>
-                    <td>100</td>
+                    <td>200</td>
                     <td>개</td>
-                    <td>2021.06.22 11:53</td>
-                    <td>2021.06.22 11:53</td>
-                    <td> 
-                      
-                      <button class="btn btn-primary btn-xs"><i class="fa fa-pencil"></i></button>
-                      
-                    </td>
-                  </tr>
+                    <td>2021-01-18 16:30</td>
+                    <td><span class="label label-warning label-mini">출고준비중</span></td>
+                    <td>2021-01-18 16:30</td> 
+                  	<td>  
+                  		<button type="submit" th:href="@{/modifyHumanResources}" class="btn btn-primary btn-xs">
+                  			<i class="fa fa-pencil"></i>
+                  				수정
+                  		</button>
+                  	</td>
+                   </tr>
                   
+              
+                 
+                 
                
-
                  
                  
                 </tbody>
@@ -118,16 +128,14 @@ layout:decorate="@{layout/default_rawMaterials}">
             
 		        <div class="form-group text-right">
 		          <div class="col-lg-offset-2 col-lg-10">
-		           <a th:href="@{/addExWarehousing}" class="btn btn-success btn-xs"><i class="fa fa-check"></i>등록</a>
+		           <a th:href="@{/addWareHousing}" class="btn btn-success btn-xs"><i class="fa fa-check"></i>등록</a>
 		           <a id="deleteBtn" class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i>삭제</a>
 		          </div>
 		        </div>
        		</form>
        		
           </div>
-          <!-- /col-md-12 -->
         </div>
-        <!-- /row -->
         
 </th:block>
