--- teamBootstrap/src/main/resources/templates/system/sendEmail.html
@@ -1,144 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-xmlns:layout="http://www.ultraq.net.nz/thymeleaf/Layout"
-layout:decorate="@{layout/default_system}">
-
-<th:block layout:fragment="pageTitle">
-	   <div class="border-head"> 
-	    <h3>이메일</h3>
-	    </div>
-</th:block>
-
-
-<th:block layout:fragment="container">
-	
-      
-        <div class="row mt">
-          <div class="col-sm-3">
-            <section class="panel">
-              <div class="panel-body">
-                <a href="mail_compose.html" class="btn btn-compose">
-                  <i class="fa fa-pencil"></i>  Compose Mail
-                  </a>
-                <ul class="nav nav-pills nav-stacked mail-nav">
-                  <li class="active"><a href="inbox.html"> <i class="fa fa-inbox"></i> Inbox  <span class="label label-theme pull-right inbox-notification">3</span></a></li>
-                  <li><a href="#"> <i class="fa fa-envelope-o"></i> Send Mail</a></li>
-                  <li><a href="#"> <i class="fa fa-exclamation-circle"></i> Important</a></li>
-                  <li><a href="#"> <i class="fa fa-file-text-o"></i> Drafts <span class="label label-info pull-right inbox-notification">8</span></a></a>
-                  </li>
-                  <li><a href="#"> <i class="fa fa-trash-o"></i> Trash</a></li>
-                </ul>
-              </div>
-            </section>
-            <section class="panel">
-              <div class="panel-body">
-                <ul class="nav nav-pills nav-stacked labels-info ">
-                  <li>
-                    <h4>Friends Online</h4>
-                  </li>
-                  <li>
-                    <a href="#">
-                        <img th:src="@{Dashio/img/friends/fr-10.jpg}" class="img-circle" width="20">Laura
-                        <p><span class="label label-success">Available</span></p>
-                      </a>
-                  </li>
-                  <li>
-                    <a href="#">
-                        <img th:src="@{Dashio/img/friends/fr-05.jpg}" class="img-circle" width="20">David
-                        <p><span class="label label-danger"> Busy</span></p>
-                      </a>
-                  </li>
-                  <li>
-                    <a href="#">
-                        <img th:src="@{Dashio/img/friends/fr-01.jpg}" class="img-circle" width="20">Mark
-                        <p>Offline</p>
-                      </a>
-                  </li>
-                  <li>
-                    <a href="#">
-                        <img th:src="@{Dashio/img/friends/fr-03.jpg}" class="img-circle" width="20">Phillip
-                        <p>Offline</p>
-                      </a>
-                  </li>
-                  <li>
-                    <a href="#">
-                        <img th:src="@{Dashio/img/friends/fr-02.jpg}" class="img-circle" width="20">Joshua
-                        <p>Offline</p>
-                      </a>
-                  </li>
-                </ul>
-                <a href="#"> + Add More</a>
-                <div class="inbox-body text-center inbox-action">
-                  <div class="btn-group">
-                    <a class="btn mini btn-default" href="javascript:;">
-                      <i class="fa fa-power-off"></i>
-                      </a>
-                  </div>
-                  <div class="btn-group">
-                    <a class="btn mini btn-default" href="javascript:;">
-                      <i class="fa fa-cog"></i>
-                      </a>
-                  </div>
-                </div>
-              </div>
-            </section>
-          </div>
-          <div class="col-sm-9">
-            <section class="panel">
-              <header class="panel-heading wht-bg">
-                <h4 class="gen-case">
-                    Compose Mail
-                    <form action="#" class="pull-right mail-src-position">
-                      <div class="input-append">
-                        <input type="text" class="form-control " placeholder="Search Mail">
-                      </div>
-                    </form>
-                  </h4>
-              </header>
-              <div class="panel-body">
-                <div class="compose-btn pull-right">
-                  <button class="btn btn-theme btn-sm"><i class="fa fa-check"></i> Send</button>
-                  <button class="btn btn-sm"><i class="fa fa-times"></i> Discard</button>
-                  <button class="btn btn-sm">Draft</button>
-                </div>
-                <div class="compose-mail">
-                  <form role="form-horizontal" method="post">
-                    <div class="form-group">
-                      <label for="to" class="">To:</label>
-                      <input type="text" tabindex="1" id="to" class="form-control">
-                      <div class="compose-options">
-                        <a onclick="$(this).hide(); $('#cc').parent().removeClass('hidden'); $('#cc').focus();" href="javascript:;">Cc</a>
-                        <a onclick="$(this).hide(); $('#bcc').parent().removeClass('hidden'); $('#bcc').focus();" href="javascript:;">Bcc</a>
-                      </div>
-                    </div>
-                    <div class="form-group hidden">
-                      <label for="cc" class="">Cc:</label>
-                      <input type="text" tabindex="2" id="cc" class="form-control">
-                    </div>
-                    <div class="form-group hidden">
-                      <label for="bcc" class="">Bcc:</label>
-                      <input type="text" tabindex="2" id="bcc" class="form-control">
-                    </div>
-                    <div class="form-group">
-                      <label for="subject" class="">Subject:</label>
-                      <input type="text" tabindex="1" id="subject" class="form-control">
-                    </div>
-                    <div class="compose-editor">
-                      <textarea class="wysihtml5 form-control" rows="9"></textarea>
-                      <input type="file" class="default">
-                    </div>
-                    <div class="compose-btn">
-                      <button class="btn btn-theme btn-sm"><i class="fa fa-check"></i> Send</button>
-                      <button class="btn btn-sm"><i class="fa fa-times"></i> Discard</button>
-                      <button class="btn btn-sm">Draft</button>
-                    </div>
-                  </form>
-                </div>
-              </div>
-            </section>
-          </div>
-        </div>
- 
-</th:block>
-
-</html>
